fungiclef-effnet/pyproject.toml
[tool.poetry]
name = "fungiclef-effnet"
version = "0.1.0"
description = ""
authors = ["Jack Etheredge <jack.etheredge@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.9"
pandas = "^2.2.1"
tqdm = "^4.66.2"
timm = "^0.9.16"
torch = {version = "^2.2.2+cu118", source = "pytorch-gpu-src"}
torchvision = {version = "^0.17.2+cu118", source = "pytorch-gpu-src"}
torchaudio = {version = "^2.2.2+cu118", source = "pytorch-gpu-src"}
matplotlib = "^3.8.3"
scikit-learn = "^1.4.1.post1"
seesawloss = "^0.1.1"
focal-loss-torch = "^0.1.2"
imbalanced-learn = "^0.12.0"
torch-lr-finder = "^0.2.1"
hydra-core = "^1.3.2"
opencv-python = "^4.9.0.80"
h5py = "^3.11.0"

[tool.poetry.dev-dependencies]

[tool.poetry.group.dev.dependencies]
setuptools = "^69.2.0"
black = "^24.3.0"


[[tool.poetry.source]]
name = "pytorch-gpu-src"
url = "https://download.pytorch.org/whl/cu118"
priority = "explicit"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
